Digital RTC calibration
STM32 microcontrollers allow digital calibration of the real time
clock frequency with a range from -487.1 to +488.5 ppm.

To facilitate calibration, it is possible to output the calibrated
oscillator output (RTC_CALIB) on pin PC13. On the 2.8" NanoVNA-H
this pin is wired to LED2 (blue led) via the 1k R3 resistor, which
is easy to probe with an oscilloscope or frequency counter.

Expose both features in the "RTC CAL" submenu under "EXPERT SETTINGS"

Signed-off-by: Nicola Di Lieto <nicola.dilieto@gmail.com>